requirements.txt
# ============================================================================
# BAAS ARC - Python Dependencies
# ============================================================================
# Autonomous Banking for AI Agents - Arc x Circle Hackathon 2026
# Compatible with Python 3.10+
# ============================================================================

# ----------------------------------------------------------------------------
# Web Framework & API
# ----------------------------------------------------------------------------
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pydantic==2.6.1

# ----------------------------------------------------------------------------
# Blockchain & Web3 (Arc Network)
# ----------------------------------------------------------------------------
web3==6.15.1
eth-account==0.11.0
eth-utils==4.0.0
eth-typing==3.5.2
hexbytes==0.3.1

# ----------------------------------------------------------------------------
# Circle & USDC Integration
# ----------------------------------------------------------------------------
circle-sdk==1.0.0  # Circle Programmable Wallets SDK

# ----------------------------------------------------------------------------
# AI & Machine Learning (Gemini)
# ----------------------------------------------------------------------------
google-generativeai==0.4.0
google-ai-generativelanguage==0.4.0

# ----------------------------------------------------------------------------
# DeFi Protocols (Aave)
# ----------------------------------------------------------------------------
aave-python==0.5.0

# ----------------------------------------------------------------------------
# Data Processing & Storage
# ----------------------------------------------------------------------------
python-dotenv==1.0.0
python-dateutil==2.8.2
pyyaml==6.0.1

# ----------------------------------------------------------------------------
# Cryptography & Security
# ----------------------------------------------------------------------------
cryptography==41.0.7
pycryptodome==3.19.0
ecdsa==0.18.0

# ----------------------------------------------------------------------------
# Testing & Quality
# ----------------------------------------------------------------------------
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# ----------------------------------------------------------------------------
# UI Dashboard
# ----------------------------------------------------------------------------
plotly==5.18.0

# ----------------------------------------------------------------------------
# Utilities
# ----------------------------------------------------------------------------
colorama==0.4.6
structlog==23.2.0
cachetools==5.3.2

# ============================================================================
# OPTIONAL DEPENDENCIES (for production deployment)
# ============================================================================

# Production Database (uncomment if using PostgreSQL)
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.23

# Redis Cache (uncomment if using Redis)
# redis==5.0.1
# hiredis==2.2.3

# Production Server (uncomment for production deployment)
# gunicorn==21.2.0

# Advanced Monitoring (uncomment if needed)
# prometheus-client==0.19.0

# ZK-Proofs (uncomment if using zero-knowledge proofs)
# circom  # Requires separate installation
# snarkjs  # Requires npm install -g snarkjs

# ============================================================================
# INSTALLATION
# ============================================================================
#
# 1. Create virtual environment:
#    python -m venv venv
#
# 2. Activate:
#    Windows: venv\Scripts\activate
#    Linux/Mac: source venv/bin/activate
#
# 3. Install:
#    pip install -r requirements.txt
#
# 4. Verify:
#    python scripts/setup.py
#
# ============================================================================
